feat: 지연로딩 기능 추가
